+++ platform/src/sdk/models/operations/listintegrations.ts
@@ -53,4 +53,6 @@ export class ListIntegrationsResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    next: () => Promise<ListIntegrationsResponse | null>;
 }